% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.all.unique.R
\name{is.all.unique}
\alias{is.all.unique}
\title{is all unique}
\usage{
is.all.unique(x)
}
\arguments{
\item{x}{is a given string}
}
\description{
This function checks string if it consists of different characters or not.
Takes a string and splits its characters.
Compares length of the characters and length of the unique characters in a given string.
Writes all unique if the string consists of unique letters.
Writes duplicates found if the string consists of same letters.
}
\examples{
x <-"leman"

is.all.unique(x)

> "all unique"

 ------------------------------------

x <-"efe"

is.all.unique(x)

> "duplicates found."

}
