% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.NOs.R
\name{missing.NOs}
\alias{missing.NOs}
\title{Finding Missing Numbers}
\usage{
missing.NOs(x)
}
\arguments{
\item{x}{is a given list of integers.}
}
\description{
This function finds missing numbers in a given list which consist of numbers between 1 to 100.
Takes a list and finds missing numbers by using setdiff function.
}
\examples{
Set an example list to test.

 x <- c(1:45,48:100)

 missing.NOs(x)

 46
 47

}
