% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.of.the.others.R
\name{product.of.the.others}
\alias{product.of.the.others}
\title{Product of the Other Values}
\usage{
product.of.the.others(x)
}
\arguments{
\item{x}{is a given vector}
}
\description{
This function products other numbers in a given vector and writes on it.
Creates a vector which consists of missing values and has the same length with a given vector.
Makes a loop for each values in a vector.
Products of the numbers other then itself.
}
\examples{
x <- c(1,2,3,4)

product.of.the.others (x)

 > 24 12  8  6

}
